+++ frontend/src/sections/history/Findings/components/ExcelPrintout.tsx
@@ -390,6 +390,9 @@ const generateSheet3 = (
       }
     })
 
+    // Insert an empty row before the note
+    sheet3.addRow({});
+
     const lastRow = sheet3?.addRow({
       no: "",
       checklist_num: `: Permasalahan telah diselesaikan (tidak dihitung sebagai permasalahan final)`,