+++ frontend/src/sections/history/Findings/components/FollowUpTable.tsx
@@ -3,11 +3,13 @@
  * © Kanwil DJPb Sumbar 2024
  */
 
+import { useState} from 'react';
 import {Stack, Typography, Table, Card, CardHeader, TableSortLabel,
-        TableHead, Grow, TableBody, TableRow, TableCell} from '@mui/material';
+        TableHead, Grow, TableBody, TableRow, TableCell, Button} from '@mui/material';
 import { useTheme} from '@mui/material/styles';
-import { FindingsResponseType } from '../../../followUp/types';
 import ExcelPrintout from './ExcelPrintout';
+import { DerivedFindingsType } from '../../../../types/findings.type';
+import Iconify from '../../../../components/iconify';
 // ---------------------------------------------------
 const TABLE_HEAD = [
   { id: 'no', label: 'No', alignRight: false },
@@ -22,32 +24,46 @@ const TABLE_HEAD = [
 ];
 
 interface FollowUpTableProps{
-  findings: FindingsResponseType[] | [],
+  isFinal: boolean | null,
+  nonFinalFindings: DerivedFindingsType[] | null,
+  finalFindings: DerivedFindingsType[] | null, 
   kppnId: string | null,
   period: number
 }
 
 // ----------------------------------------------------------------------------------
-export default function FollowUpTable({findings, kppnId, period}: FollowUpTableProps) {
+export default function FollowUpTable({isFinal, nonFinalFindings, finalFindings, kppnId, period}: FollowUpTableProps) {
   const theme = useTheme();
 
-  const refKomponen = Array(...new Set(findings.map((item) => item.matrixDetail[0].komponen_string)));
+  const [showFinal, setShowFinal] = useState<boolean>(isFinal || false);
 
-  const numberedFindings = findings.map((item, index) => ({
+  const allowFinalState = isFinal && showFinal;
+
+  const findings = allowFinalState ? finalFindings : nonFinalFindings;
+
+  const refKomponen = Array(...new Set(findings?.map((item) => item.komponen.title)));
+
+  const handleChangeFinal = (final: boolean) => {
+    if(isFinal){
+      setShowFinal(final);
+    }
+  };
+
+  const numberedFindings = findings?.map((item, index) => ({
     ...item,
     nomor: index + 1
   }));
 
   const refSubKomponen =  refKomponen.map((k) => {
-    const findingsByKomponen = findings.filter(f => f.matrixDetail[0].komponen_string === k);
-    const subkomponen = Array(...new Set(findingsByKomponen.map((item) => item.matrixDetail[0].subkomponen_string)));
+    const findingsByKomponen = findings?.filter(f => f.komponen.title === k);
+    const subkomponen = Array(...new Set(findingsByKomponen?.map((item) => item.subkomponen.title)));
 
     return ({
       komponen: k,
       subkomponen: subkomponen
     }
     )
-  })
+  });
 
   return (
     <>
@@ -57,7 +73,7 @@ export default function FollowUpTable({findings, kppnId, period}: FollowUpTableP
           title={
             <>
               <Stack direction="row"  justifyContent={'space-between'} maxWidth={'100%'}>
-                <Typography variant='h6'>Rekapitulasi Permasalahan</Typography>
+                <Typography variant='h6'>{`Rekapitulasi Permasalahan ${allowFinalState ? '(Final)' : '(Non-Final)'}`}</Typography>
                 <ExcelPrintout kppnId={kppnId || ""} period={period}/>
               </Stack>
             </>
@@ -93,42 +109,42 @@ export default function FollowUpTable({findings, kppnId, period}: FollowUpTableP
                       </TableCell>
                     </TableRow>
                     {
-                        numberedFindings.filter(f => f.matrixDetail[0].komponen_string === item.komponen).map((item) =>{
+                        numberedFindings?.filter(f => f.komponen.title === item.komponen).map((item) =>{
                           return (
                             <>
                               <TableRow hover key={item.id} tabIndex={-1}>
                                 <TableCell align="justify" sx={{fontSize: '13px'}}>{item.nomor}</TableCell>
 
                                 <TableCell align="left" sx={{fontSize: '13px'}}>
-                                  {item.matrixDetail[0].checklist[0].matrix_title}
+                                  {item.checklist.title}
                                 </TableCell>
 
                                 <TableCell align="left" sx={{fontSize: '13px'}}>
-                                  {item.matrixDetail[0].hasil_implementasi}
+                                  {item.matrix.hasil_implementasi}
                                 </TableCell>
 
                                 <TableCell align="left" sx={{fontSize: '13px'}}>
-                                  {item.matrixDetail[0].permasalahan}
+                                  {item.matrix.permasalahan}
                                 </TableCell>
 
                                 <TableCell align="left" sx={{fontSize: '13px'}}>
-                                  {item.matrixDetail[0].rekomendasi}
+                                  {item.matrix.rekomendasi}
                                 </TableCell>
 
                                 <TableCell align="left" sx={{fontSize: '13px'}}>
-                                  {item.matrixDetail[0].peraturan}
+                                  {item.matrix.peraturan}
                                 </TableCell>
 
                                 <TableCell align="left" sx={{fontSize: '13px'}}>
-                                  {item.matrixDetail[0].uic}
+                                  {item.matrix.uic}
                                 </TableCell> 
 
                                 <TableCell align="left" sx={{fontSize: '13px'}}>
-                                  {item.matrixDetail[0].ws_junction[0].kanwil_score}
+                                  {item.ws_junction.kanwil_score}
                                 </TableCell> 
 
                                 <TableCell align="left" sx={{fontSize: '13px'}}>
-                                  {item.matrixDetail[0].subkomponen_string}
+                                  {item.subkomponen.title}
                                 </TableCell> 
                               </TableRow>
                             </>  
@@ -145,8 +161,32 @@ export default function FollowUpTable({findings, kppnId, period}: FollowUpTableP
         </Table>
       </Card>
       </Grow>
-
-      <Grow in>
+      
+      {
+        isFinal ? showFinal 
+          ? (
+              <Button 
+                variant='contained' 
+                endIcon={<Iconify icon="solar:eye-bold-duotone" />} 
+                sx={{mt: 2}}
+                onClick={() => handleChangeFinal(false)}
+              >
+                Show Permasalahan Non Final
+              </Button>
+            )
+          : (
+              <Button 
+                variant='contained' 
+                endIcon={<Iconify icon="solar:eye-closed-bold-duotone" />} 
+                sx={{mt: 2}}
+                onClick={() => handleChangeFinal(true)}
+              >
+                Hide Permasalahan Non Final
+              </Button>
+            )
+        : null
+      }
+      {/* <Grow in>
         <Stack direction='column' spacing={1} sx={{pl: 2}}>
           <Typography variant='body2' fontWeight='bold' sx={{fontSize: '12px'}}>*Status:</Typography>
           <Typography variant='body2' sx={{fontSize: '12px'}}>1. Belum: belum ditindaklanjuti KPPN </Typography>
@@ -154,7 +194,7 @@ export default function FollowUpTable({findings, kppnId, period}: FollowUpTableP
           <Typography variant='body2' sx={{fontSize: '12px'}}>3. Ditolak: tindak lanjut ditolak Kanwil </Typography>
           <Typography variant='body2' sx={{fontSize: '12px'}}>4. Disetujui: tindak lanjut disetujui Kanwil </Typography>
         </Stack>
-      </Grow>
+      </Grow> */}
     </>
   )
 }