+++ frontend/src/sections/history/Info/Info.tsx
@@ -1,24 +1,65 @@
-import {Card, CardContent, Grid, Stack, Typography, Avatar, List, ListItemText} from '@mui/material';
+import {Card, CardContent, Grid, Stack, Typography} from '@mui/material';
 import useDictionary from '../../../hooks/useDictionary';
 import Iconify from '../../../components/iconify';
 import Label from '../../../components/label';
 import {useTheme} from '@mui/material/styles';
 import UserItem from './components/UserItem';
+import useAxiosJWT from '../../../hooks/useAxiosJWT';
+import { useEffect, useState } from 'react';
+import useSnackbar from '../../../hooks/display/useSnackbar';
 
 interface InfoProps{
   selectedUnit: string;
   selectedPeriod: number;
   selectedData: number;
-}
+};
 
-export default function Info({selectedUnit, selectedPeriod, selectedData}:InfoProps) {
+interface SimpleUser{
+  name: string;
+  kppn: string;
+  nip: string;
+  picture: string;
+};
+
+interface InfoData{
+  score: number | null;
+  findings: number | null;
+  isFinal: boolean;
+  openPeriod: string | null;
+  closePeriod: string | null;
+  openFollowUp: string | null;
+  closeFollowUp: string | null;
+  users: SimpleUser[] | null;
+};
+
+export default function Info({selectedUnit, selectedPeriod}:InfoProps) {
   const theme = useTheme();
+
+  const axiosJWT = useAxiosJWT();
+
+  const [info, setInfo] = useState<InfoData | null>(null);
+
+  const {openSnackbar} = useSnackbar();
   
-  const {komponenRef, periodRef, kppnRef, subKomponenRef} = useDictionary();
+  const { periodRef, kppnRef} = useDictionary();
 
   const unitString = kppnRef?.list.filter((item) => item.id === selectedUnit)?.[0]?.alias || '';
 
   const periodString = periodRef?.list.filter((item) => item.id === selectedPeriod)?.[0]?.name || '';
+
+  const getData = async() => {
+    try{
+      const response = await axiosJWT.get(`/info/getByKPPN/${selectedUnit}/${selectedPeriod}`);
+      setInfo(response.data);
+    }catch(err: any){
+      setInfo(null);
+      openSnackbar(err?.response?.data?.message, "error");
+    }
+  }
+
+  useEffect(() => {
+    getData();
+  }, [selectedUnit, selectedPeriod]);
 
   return (
     <>
@@ -38,7 +79,7 @@ export default function Info({selectedUnit, selectedPeriod, selectedData}:InfoPr
                 </div>
                 <div>
                   <Typography fontWeight={600} variant="body2">{`Nilai Kinerja KPPN :`}</Typography>
-                  <Typography variant='body3'>{`9.25`} </Typography>
+                  <Typography variant='body3'>{info?.score?.toFixed(2) || '-'} </Typography>
                 </div>
               </Stack>
               <Stack direction={'row'} gap={2} marginBottom={4}>
@@ -47,7 +88,9 @@ export default function Info({selectedUnit, selectedPeriod, selectedData}:InfoPr
                 </div>
                 <div>
                   <Typography fontWeight={600} variant="body2">{`Permasalahan :`}</Typography>
-                  <Typography variant='body3'>{`8 (Final)`} </Typography>
+                  <Typography variant='body3'>
+                    {`${info?.findings || '-'} ${(info?.isFinal? "(Final)" : "(Non Final)")}`}
+                  </Typography>
                 </div>
               </Stack>
               <Stack direction={'row'} gap={2} marginBottom={4}>
@@ -56,9 +99,9 @@ export default function Info({selectedUnit, selectedPeriod, selectedData}:InfoPr
                 </div>
                 <div>
                   <Typography fontWeight={600} variant="body2">{`Periode Pembinaan :`}</Typography>
-                  <Label color='success'>12/03/2023</Label>
+                  <Label color='success'>{new Date(info?.openPeriod || '').toLocaleDateString('id-ID')}</Label>
                   <Typography variant='body3'>{`-`} </Typography>
-                  <Label color='error'>01/05/2024</Label>
+                  <Label color='error'>{new Date(info?.closePeriod || '').toLocaleDateString('id-ID')}</Label>
                 </div>
               </Stack>
               <Stack direction={'row'} gap={2} marginBottom={4}>
@@ -67,9 +110,9 @@ export default function Info({selectedUnit, selectedPeriod, selectedData}:InfoPr
                 </div>
                 <div>
                   <Typography fontWeight={600} variant="body2">{`Periode Tindak Lanjut :`}</Typography>
-                  <Label color='success'>12/03/2023</Label>
+                  <Label color='success'>{new Date(info?.openFollowUp || '').toLocaleDateString('id-ID')}</Label>
                   <Typography variant='body3'>{`-`} </Typography>
-                  <Label color='error'>01/05/2024</Label>
+                  <Label color='error'>{new Date(info?.closeFollowUp || '').toLocaleDateString('id-ID')}</Label>
                 </div>
               </Stack>
               <Stack direction={'row'} gap={2} marginBottom={4}>
@@ -80,9 +123,9 @@ export default function Info({selectedUnit, selectedPeriod, selectedData}:InfoPr
                   <Typography fontWeight={600} variant="body2">{`User Kanwil :`}</Typography>
                   <div style={{ display: 'flex', flexWrap: 'wrap', gap:'1rem'}}>
                     {
-                      new Array(10).fill(10).map((item, index) => (
+                      info?.users?.filter((item) => item.kppn.length === 5).map((item, index) => (
                         <>
-                          <UserItem name="" nip="/" />
+                          <UserItem name={item.name} nip={item.nip} key={index} avatar={item.picture}/>
                         </>
                       ))
                     }
@@ -97,9 +140,9 @@ export default function Info({selectedUnit, selectedPeriod, selectedData}:InfoPr
                   <Typography fontWeight={600} variant="body2">{`User KPPN :`}</Typography>
                   <div style={{ display: 'flex', flexWrap: 'wrap', gap:'1rem'}}>
                     {
-                      new Array(10).fill(10).map((item, index) => (
+                      info?.users?.filter((item) => item.kppn.length !== 5).map((item, index) => (
                         <>
-                          <UserItem name="" nip="/" />
+                          <UserItem name={item.name} nip={item.nip} key={index} avatar={item.picture}/>
                         </>
                       ))
                     }