+++ frontend/src/sections/history/Standardization/Standardization.tsx
@@ -8,8 +8,9 @@ import StandardizationTable from "./components/StandardizationTable";
 import useStandardization from "./useStandardization";
 import usePreviewFileModal from "./usePreviewFileModal";
 import PreviewFileModal from "./components/PreviewFileModal";
-import { Stack } from "@mui/material";
+import { Stack, Card, CardContent, Typography } from "@mui/material";
 import PageLoading from "../../../components/pageLoading";
+import useDictionary from "../../../hooks/useDictionary";
 // --------------------------------------------------------------
 
 interface StandardizationProps{
@@ -24,6 +25,12 @@ export default function Standardization({selectedUnit, selectedPeriod, selectedD
   const {open, modalOpen, modalClose, file} = usePreviewFileModal();
 
   const {getStandardization} = useStandardization();
+
+    const { periodRef, kppnRef } = useDictionary();
+
+  const unitString = kppnRef?.list.filter((item) => item.id === selectedUnit)?.[0]?.alias || '';
+
+  const periodString = periodRef?.list.filter((item) => item.id === selectedPeriod)?.[0]?.name || '';
 
   useEffect(() => {
     async function getData(){
@@ -42,32 +49,40 @@ export default function Standardization({selectedUnit, selectedPeriod, selectedD
 
   return (
     <>
-       <Stack direction='column' spacing={4}>
-          {loading
-            ?<PageLoading duration={2}/>
-            : <>
-                <StandardizationTable 
-                  header={'Manajemen Eksternal'} 
-                  modalOpen={modalOpen} 
-                  kppnTab={selectedUnit}
-                  cluster={1} 
-                />
-                <StandardizationTable 
-                  header={'Penguatan Kapasitas Perbendaharaan'} 
-                  modalOpen={modalOpen} 
-                  kppnTab={selectedUnit} 
-                  cluster={2} 
-                />
-                <StandardizationTable 
-                  header={'Penguatan Manajemen Internal'} 
-                  modalOpen={modalOpen} 
-                  kppnTab={selectedUnit}
-                  cluster={3}  
-                />
-              </>
-          }
-        </Stack>
-        <PreviewFileModal open={open} modalClose={modalClose} file={file} kppnId={selectedUnit} />
+      <Card>
+        <CardContent>
+          <Stack alignContent={'center'} textAlign={'center'}>
+            <Typography variant='h6'>{`Data Standardisasi`} </Typography>
+            <Typography variant='body3'>{`${unitString}, Periode ${periodString}`} </Typography>
+          </Stack>
+        </CardContent>
+      </Card>
+      <Stack direction='column' spacing={4} marginTop={4}>
+        {loading
+          ?<PageLoading duration={2}/>
+          : <>
+              <StandardizationTable 
+                header={'Manajemen Eksternal'} 
+                modalOpen={modalOpen} 
+                kppnTab={selectedUnit}
+                cluster={1} 
+              />
+              <StandardizationTable 
+                header={'Penguatan Kapasitas Perbendaharaan'} 
+                modalOpen={modalOpen} 
+                kppnTab={selectedUnit} 
+                cluster={2} 
+              />
+              <StandardizationTable 
+                header={'Penguatan Manajemen Internal'} 
+                modalOpen={modalOpen} 
+                kppnTab={selectedUnit}
+                cluster={3}  
+              />
+            </>
+        }
+      </Stack>
+      <PreviewFileModal open={open} modalClose={modalClose} file={file} kppnId={selectedUnit} />
     </>
   )
 }