+++ frontend/src/types/matrix.type.ts
@@ -1,0 +1,13 @@
+export interface MatrixType{
+  id: number,
+  ws_junction_id: number,
+  worksheet_id: string,
+  checklist_id: number,
+  hasil_implementasi: string,
+  permasalahan: string, 
+  rekomendasi: string,
+  peraturan: string,
+  uic: string,
+  tindak_lanjut: string, 
+  is_finding: number
+};