+++ frontend/src/types/opsi.type.ts
@@ -1,0 +1,9 @@
+export interface OpsiType{
+  id: number,
+  title: string, 
+  value: number,
+  checklist_id: number,
+  positive_fallback: string,
+  negative_fallback: string,
+  rekomendasi: string
+};